changing variables related to the number of entries stored for an event from int64_t to int32_t

+++ libhpx/instrumentation/prof_basic.c
@@ -71,7 +71,7 @@ static void _create_new_list(profile_list_t new_list, char *key, bool simple){
   _profile_log.events[index].simple = simple;
 }
 
-static int64_t _create_new_entry(struct profile_entry new_entry,
+static int32_t _create_new_entry(struct profile_entry new_entry,
                               int event, bool simple){
   if(_profile_log.events[event].num_entries ==
      _profile_log.events[event].max_entries){
@@ -79,7 +79,7 @@ static int64_t _create_new_entry(struct profile_entry new_entry,
     struct profile_entry *new_list =
                                 malloc(_profile_log.events[event].max_entries *
                                        sizeof(struct profile_entry));
-    for(int64_t i = 0; i < _profile_log.events[event].num_entries; i++){
+    for(int32_t i = 0; i < _profile_log.events[event].num_entries; i++){
       new_list[i].start_time = _profile_log.events[event].entries[i].start_time; 
       new_list[i].run_time = _profile_log.events[event].entries[i].run_time;     
       new_list[i].counter_totals = _profile_log.events[event].entries[i].counter_totals;
@@ -92,7 +92,7 @@ static int64_t _create_new_entry(struct profile_entry new_entry,
     _profile_log.events[event].entries = new_list;
   }
 
-  int64_t index = _profile_log.events[event].num_entries;
+  int32_t index = _profile_log.events[event].num_entries;
   _profile_log.events[event].tally++;
   _profile_log.events[event].num_entries++;
   _profile_log.events[event].entries[index] = new_entry;
@@ -158,7 +158,7 @@ int prof_get_maximums(int64_t *values, char *key){
   return LIBHPX_OK;
 }
 
-int64_t prof_get_tally(char *key){
+int32_t prof_get_tally(char *key){
   int event = _get_event_num(key);
   if(event == HPX_PROF_NO_RESULT){
     return 0;
@@ -173,7 +173,7 @@ void prof_get_average_time(char *key, hpx_time_t *avg){
   }
 
   uint64_t seconds, ns, average = 0;
-  for(int64_t i = 0; i < _profile_log.events[event].num_entries; i++){
+  for(int32_t i = 0; i < _profile_log.events[event].num_entries; i++){
     if(_profile_log.events[event].entries[i].marked){
       average += hpx_time_diff_ns(TIME_NULL,
                                   _profile_log.events[event].entries[i].run_time);
@@ -196,7 +196,7 @@ void prof_get_total_time(char *key, hpx_time_t *tot){
   }
 
   uint64_t seconds, ns, total = 0;
-  for(int64_t i = 0; i < _profile_log.events[event].num_entries; i++){
+  for(int32_t i = 0; i < _profile_log.events[event].num_entries; i++){
     if(_profile_log.events[event].entries[i].marked){
       total += hpx_time_diff_ns(TIME_NULL,
                                 _profile_log.events[event].entries[i].run_time);
@@ -216,10 +216,10 @@ void prof_get_min_time(char *key, hpx_time_t *min){
 
   int64_t seconds, ns, temp;
   int64_t minimum = 0;
-  int64_t start = _profile_log.events[event].num_entries;
+  int32_t start = _profile_log.events[event].num_entries;
 
   if(_profile_log.events[event].num_entries > 0 ){
-    for(int64_t i = 0; i < _profile_log.events[event].num_entries; i++){
+    for(int32_t i = 0; i < _profile_log.events[event].num_entries; i++){
       if(_profile_log.events[event].entries[i].marked){
         minimum = 
              hpx_time_diff_ns(TIME_NULL,
@@ -229,7 +229,7 @@ void prof_get_min_time(char *key, hpx_time_t *min){
       }
     }
   }
-  for(int64_t i = start; i < _profile_log.events[event].num_entries; i++){
+  for(int32_t i = start; i < _profile_log.events[event].num_entries; i++){
     if(_profile_log.events[event].entries[i].marked){
       temp = hpx_time_diff_ns(TIME_NULL,
                               _profile_log.events[event].entries[i].run_time);
@@ -252,10 +252,10 @@ void prof_get_max_time(char *key, hpx_time_t *max){
 
   int64_t seconds, ns, temp;
   int64_t maximum = 0;
-  int64_t start = _profile_log.events[event].num_entries;
+  int32_t start = _profile_log.events[event].num_entries;
 
   if(_profile_log.events[event].num_entries > 0 ){
-    for(int64_t i = 0; i < _profile_log.events[event].num_entries; i++){
+    for(int32_t i = 0; i < _profile_log.events[event].num_entries; i++){
       if(_profile_log.events[event].entries[i].marked){
         maximum = 
              hpx_time_diff_ns(TIME_NULL,
@@ -265,7 +265,7 @@ void prof_get_max_time(char *key, hpx_time_t *max){
       }
     }
   }
-  for(int64_t i = start; i < _profile_log.events[event].num_entries; i++){
+  for(int32_t i = start; i < _profile_log.events[event].num_entries; i++){
     if(_profile_log.events[event].entries[i].marked){
       temp = hpx_time_diff_ns(TIME_NULL,
                               _profile_log.events[event].entries[i].run_time);
@@ -329,7 +329,7 @@ void prof_start_timing(char *key, int *tag){
   }
 
   struct profile_entry entry;
-  int64_t index = _create_new_entry(entry, event, _profile_log.events[event].simple);
+  int32_t index = _create_new_entry(entry, event, _profile_log.events[event].simple);
   _profile_log.events[event].entries[index].last_entry = _profile_log.current_entry;
   _profile_log.events[event].entries[index].last_event = _profile_log.current_event;
   _profile_log.current_entry = index;
@@ -345,7 +345,7 @@ int prof_stop_timing(char *key, int *tag){
   }
  
   if(*tag == HPX_PROF_NO_TAG){
-    for(int64_t i = _profile_log.events[event].num_entries - 1; i >= 0; i--){
+    for(int32_t i = _profile_log.events[event].num_entries - 1; i >= 0; i--){
       if(!_profile_log.events[event].entries[i].marked){
         *tag = i;
         break;
@@ -396,7 +396,7 @@ int prof_pause(char *key, int *tag){
   }
 
   if(*tag == HPX_PROF_NO_TAG){
-    for(int64_t i = _profile_log.events[event].num_entries - 1; i >= 0; i--){
+    for(int32_t i = _profile_log.events[event].num_entries - 1; i >= 0; i--){
       if(!_profile_log.events[event].entries[i].marked && 
          !_profile_log.events[event].entries[i].paused){
         *tag = i;
@@ -428,7 +428,7 @@ int prof_resume(char *key, int *tag){
   }
 
   if(*tag == HPX_PROF_NO_TAG){
-    for(int64_t i = _profile_log.events[event].num_entries - 1; i >= 0; i--){
+    for(int32_t i = _profile_log.events[event].num_entries - 1; i >= 0; i--){
       if(!_profile_log.events[event].entries[i].marked && 
          _profile_log.events[event].entries[i].paused){
         *tag = i;
